added chartList
