+++ src/parts/TheHeader.vue
@@ -1,5 +1,13 @@
 <template>
   <header class="bg-[#f9fafb] p-4 text-center">
-    <slot>タイトル</slot>
+    <h2 class="font-bold text-xl">{{ title }}</h2>
   </header>
 </template>
+
+<script setup>
+  import { defineProps } from "vue";
+
+  defineProps({
+    title: String
+  })
+</script>
